Modularizing files and added tqdm for training
