+++ src/test.py
@@ -1,4 +1,7 @@
 from utils import *
+from dataloader import *
+from model import *
+from parser import *
 import torch
 import random
 import torchvision
